+++ src/main/java/com/nhnacademy/bookstoreorderapi/order/client/user/dto/request/PointType.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.bookstoreorderapi.payment.dto.Request;
+package com.nhnacademy.bookstoreorderapi.order.client.user.dto.request;
 
 import lombok.Getter;
