+++ src/main/java/com/nhnacademy/bookstoreorderapi/payment/service/impl/PaymentServiceImpl.java
@@ -1,7 +1,6 @@
 package com.nhnacademy.bookstoreorderapi.payment.service.impl;
 
-import com.nhnacademy.bookstoreorderapi.order.client.user.UserServiceClient;
-import com.nhnacademy.bookstoreorderapi.order.client.user.dto.ResponsePointType;
+import com.nhnacademy.bookstoreorderapi.common.service.PointService;
 import com.nhnacademy.bookstoreorderapi.order.domain.Order;
 import com.nhnacademy.bookstoreorderapi.order.domain.OrderStatus;
 import com.nhnacademy.bookstoreorderapi.order.exception.notfound.OrderNotFoundException;
@@ -11,7 +10,9 @@ import com.nhnacademy.bookstoreorderapi.payment.config.TossPaymentConfig;
 import com.nhnacademy.bookstoreorderapi.payment.domain.PayType;
 import com.nhnacademy.bookstoreorderapi.payment.domain.PaymentStatus;
 import com.nhnacademy.bookstoreorderapi.payment.domain.entity.Payment;
-import com.nhnacademy.bookstoreorderapi.payment.dto.Request.*;
+import com.nhnacademy.bookstoreorderapi.payment.dto.Request.CancelPaymentRequest;
+import com.nhnacademy.bookstoreorderapi.payment.dto.Request.PaymentApprovalRequestDto;
+import com.nhnacademy.bookstoreorderapi.payment.dto.Request.PaymentReqDto;
 import com.nhnacademy.bookstoreorderapi.payment.dto.Response.PaymentResDto;
 import com.nhnacademy.bookstoreorderapi.payment.exception.*;
 import com.nhnacademy.bookstoreorderapi.payment.repository.PaymentRepository;
@@ -36,7 +37,7 @@ public class PaymentServiceImpl implements PaymentService {
     private final PaymentRepository payRepo;
     private final TossPaymentConfig tossProps;
     private final TossPaymentClient tossClient;
-    private final UserServiceClient userServiceClient;
+    private final PointService pointService;
 
     private String extractRedirectUrl(Map<String, Object> resp) {
         return Stream.of(
@@ -67,18 +68,7 @@ public class PaymentServiceImpl implements PaymentService {
                 .filter(p -> p.getPaymentStatus() == PaymentStatus.SUCCESS)
                 .ifPresent(p -> { throw new AlreadyPaidException(orderNumber); });
 
-        // 포인트 유효성 검사
-        if (dto.getUsedPoint() > 0) {
-            Long userNo = order.getUserNo();
-            if (userNo == null) { // 비회원이 포인트를 사용하려는 경우
-                throw new NonMemberPointUsageAttemptException("비회원은 포인트를 사용할 수 없습니다.");
-            }
-            // 회원인 경우에만 포인트 검사
-            Integer currentUserPoint = userServiceClient.getUserPointByUserNo(userNo);
-            if (currentUserPoint == null || currentUserPoint < dto.getUsedPoint()) {
-                throw new NotEnoughPointException("사용하려는 포인트가 부족합니다.");
-            }
-        }
+        pointService.validatePointUsage(order.getUserNo(), dto.getUsedPoint());
 
         Map<String, Object> body = Map.of(
                 "method", dto.getPayType() == PayType.ACCOUNT ? "VIRTUAL_ACCOUNT" : dto.getPayType().name(),
@@ -147,32 +137,8 @@ public class PaymentServiceImpl implements PaymentService {
         order.setStatus(OrderStatus.PENDING_PAY);
         orderRepo.save(order);
 
-        // 결제 금액에 대한 포인트 적립 로직
-        Long userNo = order.getUserNo();
-        if (userNo != null) {
-            ResponsePointType earningRateResponse = userServiceClient.getEarningRateByUserNo(userNo);
-            int earningRate = earningRateResponse.getEarningRate();
-            int earnedPoint = (int) (payment.getPayAmount() * (earningRate / 100.0));
-
-            if (earnedPoint > 0) {
-                OrderPointPlusProcessRequest pointPlusRequest = new OrderPointPlusProcessRequest(
-                        order.getId(),
-                        earnedPoint,
-                        PointType.ORDER
-                );
-                userServiceClient.orderPointPlusProcess(userNo, pointPlusRequest);
-            }
-        }
-
-        // 포인트 차감 로직
-        if (payment.getUsedPoint() > 0) {
-            OrderPointMinusProcessRequest pointMinusRequest = new OrderPointMinusProcessRequest(
-                    order.getId(),
-                    payment.getUsedPoint(),
-                    PointType.ORDER
-            );
-            userServiceClient.orderPointMinusProcess(userNo, pointMinusRequest);
-        }
+        pointService.processEarnedPoints(order, payment);
+        pointService.processUsedPoints(order, payment);
 
         return confirmResp;
     }
